[Dataset] refactor utils module into utils folder
